requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
cloudinary==1.40.0
crispy-bootstrap4==2024.1
Django==4.2.13
django-appconf==1.0.6
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.2
django-js-asset==2.2.0
django-libsass==0.9
django-sass-processor==1.4.1
django_compressor==4.5
idna==3.7
libsass==0.23.0
packaging==24.1
pillow==10.3.0
pipdeptree==2.23.0
pkg_resources==0.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
